# openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/user-profile": {
      "get": {
        "tags": [
          "User Profile Controller"
        ],
        "summary": "Получить профиль пользователя",
        "description": "Возвращает профиль текущего авторизованного пользователя.",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "Профиль успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User Profile Controller"
        ],
        "summary": "Обновить профиль пользователя",
        "description": "Обновляет имя и аватар текущего пользователя.",
        "operationId": "updateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Профиль успешно обновлён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные профиля",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shops/{id}": {
      "get": {
        "tags": [
          "Shop Controller"
        ],
        "summary": "Получить магазин по ID",
        "description": "Возвращает магазин по указанному идентификатору.",
        "operationId": "getShop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Магазин найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "404": {
            "description": "Магазин не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shop Controller"
        ],
        "summary": "Обновить магазин",
        "description": "Обновляет информацию о магазине. Доступно только администраторам.",
        "operationId": "updateShop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Магазин успешно обновлён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "404": {
            "description": "Магазин не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Shop Controller"
        ],
        "summary": "Удалить магазин",
        "description": "Удаляет магазин по указанному идентификатору. Только для администраторов.",
        "operationId": "deleteShop",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Магазин успешно удалён"
          },
          "404": {
            "description": "Магазин не найден"
          },
          "403": {
            "description": "Доступ запрещён"
          }
        }
      }
    },
...
    "/api/v1/shops": {
      "get": {
        "tags": [
          "Shop Controller"
        ],
        "summary": "Получить список магазинов",
        "description": "Возвращает постраничный список всех магазинов.",
        "operationId": "listShops",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список магазинов получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageableResponseShopDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shop Controller"
        ],
        "summary": "Создать магазин",
        "description": "Создаёт новый магазин. Доступно только администраторам.",
        "operationId": "createShop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Магазин успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ShopDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      },
      "ShopDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "info": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "PageableResponseShopDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}